Fix Tiered sharing concept (#10)
* Fix Tiered sharing concept

.3, .7 and 1.0 of the mn block reward would end up with 2.1x the masternode share going to masternodes.

Best to break share up same as PoW.  Logically. PoW masternode share is 27% (3 + 9 + 15).  So taking 27% of the blockReward, dividing by 9 and divying up you get the same result (27/9 = 3; 3 * 1 = 3, 3*3 = 9, 3*5 = 15).

However if PoS; we use the seesaw algo to get the masternode share on a variable percent.  Then divy up same as PoW.

* fix get seesaw