Autocombine frequency (#12)
Added "oneshot" functionality, allowing you to specify a block frequency of "zero", which will run the autocombine on the processing of the next block. This functionality assumes that someone would be running their wallet infrequently, and therefore is likely to want the one shot to run when they start the wallet again. As such, the "on" and "frequency 0" is saved in the database, but after it runs, it will be shut off for the duration of the wallet run.

If they do not want it running on startup; they would simply "autocombinerewards false" after it's done it's one shot.

The feature, in it's entirety:

Defaults to off
When "on", requires a threshold (in coin count)
Block frequency defaults to every 15 blocks. This can changed to every block (resource intensive) or however many blocks one expects their threshold to hit.
Block frequency of "0" will run the sweep on the next block, and then turn it off until they run it again, or restart their wallet.
To prevent the one shot to run on startup, they simply need to turn it off after the sweep is complete (when getautocombineinfo returns "on startup" rather than "on next block")
For example:
Suppose a user has 16 tier 1 masternodes. They can expect to run into problems sending more than 75 coins (630 UTXOs at 0.12 coins per UTXO = 75.6). This would hit when each of the 16 masternodes has received 4.725 coins (75.6 / 16); which happens after about 39 block rewards (4.725 coins / 0.12 coins per reward = 39.375). If there are 35 tier one masternodes; you would expect one block every 35 blocks. 35 * 39 = 1378 blocks until you would expect to hit your 4.725 coins across 630 UTXOs.

So, a good setting would be
autocombinerewards true 4 700

Which will scan your wallet every 700 blocks and combine the dust into 4 coin transactions (or smaller if you can receive zero fee transactions.