SeeSaw Updates & Drift changes (#17)
* Change nMasternodeCountDrift to nMasternodePercentDrift

* Change MasternodeCountDrift() to MasternodePercentDrift

* Add drift boolean to GetMasternodePayment() (defaulting to false)

* Add drift calculation in if requested

* Include request for drift numbers to GetMasternodePayment()

* tweaked variable name for readability

* Update main.cpp

* debugging staking algo

* Fix SeeSaw and Drift cleanup

Fix the casting of seesaw algo to prevent flooring to zero in the decimal math.  Cleanup most of the drift code.  Still setup for testing the algo during PoW stage, so a little cleanup is needed still.

* Fixups of GetSeeSaw with drift

* Review Cleanup

A couple style nits; plus contain mRawLocked inside if (bDrift), since it's only for debugging and not needed for non-drift calculated callers.

* oops

* Fixup Dev and Fund deductions

Need to deduct the Dev and Fund payments before going into SeeSaw, so the miner isn't the one that has to pay the payments from their share (and we don't end up with a negative balance to the miner... which would be devastating; since it would end up being a very very large number to the miner since it's unsigned)

* Account for Dev and Fund payments

Need to account for Dev and Fund payments coming off the top, before the seesaw.