Posfixes (#25)
* first idea

* make vtx[0] empty and delete unused nheight

* fix nHeight

* add ntime

* add debug code

* Posfixes cs review (#20)

* Modify proof of stake resize

* Update miner.cpp

* Update wallet.cpp

Don't want to continue if nBalance is zero, as we're not staking zerocoins, and don't know what happens if we have a really big number going into SelectStakeCoins

* Fixes for Pos Blocks

* remove blob

* missing argument

* Mining Cleanup (#22)

* Modify proof of stake resize

* Update miner.cpp

* Update wallet.cpp

Don't want to continue if nBalance is zero, as we're not staking zerocoins, and don't know what happens if we have a really big number going into SelectStakeCoins

* Minor ENABLE_WALLET cleanup

To prevent compilation errors when building without ENABLE_WALLET

* Abandon CreateNewBlockWithKey() if creation during incorrect phase

CreateNewBlockWithKey() is called from both the Miner and Minter threads, but doesn't abort the block creation until way deep into the creation.  Don't bother wasting the time and resources; and abort earlier.

* Some more bitcoin miner cleanup

* Add BlockValue to vout[2] when splitting a stake

This will prevent the potential for payouts to deduct into a negative vout[2].nValue

* Remove added IncrementExtraNonce

Not sure how that works with Pivx; but IncrementExtraNonce is messing around with vtx[0].vin[0], the coinbase transaction... when there shouldn't be a coinbase transaction for PoS.  This could be a current roadblock.

* Get the right Masternode Payment

Need to calculate the masternode payment for the block height for this new block, not the previous block.  Otherwise we will fail consensus on switchover, as the seesaw won't be balanced correctly.

* Debugging Additions

Additional debugging added (plus a cleanup)

* Fix build problem

Be a little more noisy since getting the block height from a tx is a little more complex and not worth bothering with

* Clean up merge incompatibilities

Cleaned up the merge incompatibilities, and removed the fee addition (for now), until the full validation of the fee can be debated and discussed (+ consensus changes, + masternode drift accommodation)

* Cleanup Debug messages that aren't needed anymore

... and some further optimization.  And please let's not cover up the problem by not testing the block validity.  If it doesn't work here, then it shouldn't work on the consensus side either.  We don't want to be submitting blocks that we know will be rejected.

* fix typo

* fix build problem

* Update main.cpp

* Update main.cpp

* Slow down retries on PoW mine in PoS phase

Rather than shutdown the PoW mining threads on acceptance of the last PoW block, we will just slow down the attempts to keep trying the block; and keep the miner threads running into 6 confirmations of the last PoW block.  This way if somehow the chain is rewound back to proof of work, we're still working.  It's debatable if we want to keep working if we accepted a block that was ultimately rejected; but this is better than a block being rejected and all the miner wallets have already shut down their threads.

* Add scriptSig of coinbase tx to PoS block

and [readability] move scriptsig and block_value to PoW section so it is easier to read the code.

* new testnet genesis block

* Add more debugging and change GetNextWorkRequired to have correct starting point for PoS

* Debug cleanup and tweak

* Don't assert if someone else has beat us to the block

* new genesis block

* remove hash debug msg

* More debug cleanup