Revert MAX_STANDARD_TX_SIZE (#32)
Due to the consensus checks on MAX_STANDARD_TX_SIZE; increasing it will allow the current software to create the transaction, but only those running the current software will process that transaction.  So the transaction will create a network split between old and new.  To do it right, it needs to be a consensus change on a spork that is invoked when enough of the network has upgraded.  We will revisit this later, correctly, if we feel it truly is necessary.